feat(graph): rename AgentEnum to FlowAgentEnum and update references
